Fix MapDirection, rename RecruiterCreation
