+++ src/screens/index.js
@@ -15,4 +15,3 @@ export {ServiceTicket} from './Recruiter/ServiceTicket';
 export {RecruiterCreation} from './Recruiter/RecruiterCreation';
 export {CreationSecondary} from './Provider/CreationSecondary';
 export {EditProvider} from './Provider/EditProvider';
-export {MapDirection} from './Recruiter/MapDirection';
